refactor(admin): move bulk extend trial into KiloClaw subtab, add paste input
Move the standalone /admin/extend-claw-trial page into the /admin/kiloclaw
dashboard as a "Bulk Extend" tab. Add a paste-based email input mode (default)
alongside the existing CSV upload. Remove the separate sidebar nav entry.
